+++ controllers/contacts/addContact.js
@@ -1,4 +1,4 @@
-const { Contact } = require("../models/contact");
+const { Contact } = require("../../models/contact");
 
 const addContact = async (req, res) => {
   const { _id: owner } = req.user;
