+++ controllers/contacts/index.js
@@ -1,4 +1,4 @@
-const { ctrlWrapper } = require("../helpers");
+const { ctrlWrapper } = require("../../helpers");
 
 const listContacts = require("./listContacts");
 const getById = require("./getById");
