+++ controllers/contacts/removeContact.js
@@ -1,5 +1,5 @@
-const { Contact } = require("../models/contact");
-const { HttpError } = require("../helpers");
+const { Contact } = require("../../models/contact");
+const { HttpError } = require("../../helpers");
 
 const removeContact = async (req, res) => {
   const { id } = req.params;
