+++ controllers/contacts/updateContact.js
@@ -1,5 +1,5 @@
-const { Contact } = require("../models/contact");
-const { HttpError } = require("../helpers");
+const { Contact } = require("../../models/contact");
+const { HttpError } = require("../../helpers");
 
 const updateContact = async (req, res) => {
   const { id } = req.params;
