+++ controllers/users/currentUser.js
@@ -1,0 +1,8 @@
+const { User } = require("../../models/user");
+
+const currentUser = async (req, res) => {
+  const { email, subscription } = req.user;
+  res.status(200).json({ email, subscription });
+};
+
+module.exports = currentUser;
