+++ controllers/users/loginUser.js
@@ -1,0 +1,30 @@
+const jwt = require("jsonwebtoken");
+require("dotenv").config();
+
+const { User } = require("../../models/user");
+const { HttpError } = require("../../helpers");
+const { SECRET_KEY } = process.env;
+
+const loginUser = async (req, res, next) => {
+  const { email, password } = req.body;
+  const user = await User.findOne({ email });
+  if (!user || !user.validPassword(password)) {
+    throw HttpError(401, "Email or password is wrong");
+  }
+
+  const payload = {
+    id: user.id,
+  };
+
+  const token = jwt.sign(payload, SECRET_KEY, { expiresIn: "24h" });
+  await User.findByIdAndUpdate(user._id, { token });
+  res.json({
+    token,
+    user: {
+      email: user.email,
+      subscription: user.subscription,
+    },
+  });
+};
+
+module.exports = loginUser;
